python-vector-service/requirements.txt
fastapi>=0.104.0
pydantic>=2.0
uvicorn[standard]
python-multipart
pymilvus
sentence-transformers
numpy
Pillow
